Move Home to views
